+++ crates/effigy-cli/src/command_parsing_changelog.rs
@@ -1,6 +1,7 @@
 use std::path::PathBuf;
 
 use crate::value_parsing::next_required_value;
+use crate::value_parsing::parse_repo_path;
 use crate::{ChangelogArgs, ChangelogSubcommand, CliParseError, Command, HelpTopic};
 
 use super::unknown_argument;
@@ -29,11 +30,14 @@ where
     I: IntoIterator<Item = String>,
 {
     let mut file: Option<PathBuf> = None;
+    let mut repo_override: Option<PathBuf> = None;
     let mut output_json = false;
 
-    for arg in args {
+    let mut args = args.into_iter();
+    while let Some(arg) = args.next() {
         match arg.as_str() {
             "--json" => output_json = true,
+            "--repo" => repo_override = Some(parse_repo_path(&mut args)?),
             "--help" | "-h" => return Ok(Command::Help(HelpTopic::Changelog)),
             other if other.starts_with('-') => return Err(unknown_argument(other)),
             _ => file = Some(PathBuf::from(arg)),
@@ -43,6 +47,7 @@ where
     Ok(Command::Changelog(ChangelogArgs {
         subcommand: ChangelogSubcommand::Validate,
         file,
+        repo_override,
         output_json,
     }))
 }
@@ -52,12 +57,15 @@ where
     I: IntoIterator<Item = String>,
 {
     let mut file: Option<PathBuf> = None;
+    let mut repo_override: Option<PathBuf> = None;
     let mut output_json = false;
     let mut write = false;
 
-    for arg in args {
+    let mut args = args.into_iter();
+    while let Some(arg) = args.next() {
         match arg.as_str() {
             "--json" => output_json = true,
+            "--repo" => repo_override = Some(parse_repo_path(&mut args)?),
             "--write" => write = true,
             "--preview" => write = false,
             "--help" | "-h" => return Ok(Command::Help(HelpTopic::Changelog)),
@@ -69,6 +77,7 @@ where
     Ok(Command::Changelog(ChangelogArgs {
         subcommand: ChangelogSubcommand::Format { write },
         file,
+        repo_override,
         output_json,
     }))
 }
@@ -78,11 +87,14 @@ where
     I: IntoIterator<Item = String>,
 {
     let mut file: Option<PathBuf> = None;
+    let mut repo_override: Option<PathBuf> = None;
     let mut output_json = false;
 
-    for arg in args {
+    let mut args = args.into_iter();
+    while let Some(arg) = args.next() {
         match arg.as_str() {
             "--json" => output_json = true,
+            "--repo" => repo_override = Some(parse_repo_path(&mut args)?),
             "--help" | "-h" => return Ok(Command::Help(HelpTopic::Changelog)),
             other if other.starts_with('-') => return Err(unknown_argument(other)),
             _ => file = Some(PathBuf::from(arg)),
@@ -92,6 +104,7 @@ where
     Ok(Command::Changelog(ChangelogArgs {
         subcommand: ChangelogSubcommand::Analyze,
         file,
+        repo_override,
         output_json,
     }))
 }
@@ -102,12 +115,14 @@ where
 {
     let mut args = args.into_iter();
     let mut file: Option<PathBuf> = None;
+    let mut repo_override: Option<PathBuf> = None;
     let mut output_json = false;
     let mut version: Option<String> = None;
 
     while let Some(arg) = args.next() {
         match arg.as_str() {
             "--json" => output_json = true,
+            "--repo" => repo_override = Some(parse_repo_path(&mut args)?),
             "--version" => {
                 version = Some(next_required_value(
                     &mut args,
@@ -129,6 +144,7 @@ where
     Ok(Command::Changelog(ChangelogArgs {
         subcommand: ChangelogSubcommand::Extract { version },
         file,
+        repo_override,
         output_json,
     }))
 }